+++ cxx/ip/ip/shift.h
@@ -41,14 +41,16 @@ namespace bob {
         *   dimensions, and throws an exception if the shifted area and the
         *   input source array/iamge have no common points.
         * @param shift_y The y-offset of the top left corner of the shifted
-        * area wrt. to the y-index of the top left corner of the blitz::array.
+        *   area wrt. to the y-index of the top left corner of the 
+        *   blitz::array.
         * @param shift_x The x-offset of the top left corner of the shifted 
-        * area wrt. to the x-index of the top left corner of the blitz::array.
+        *   area wrt. to the x-index of the top left corner of the 
+        *   blitz::array.
         * @param src_height The height of the input image
         * @param src_width The width of the input image
         */
       void shiftParameterCheck( const int shift_y, const int shift_x,
-        const int src_height, const int src_width);
+        const size_t src_height, const size_t src_width);
     }
 
 
@@ -58,20 +60,23 @@ namespace bob {
       *   one is the width (x-axis).
       * @param src The input blitz array
       * @param dst The output blitz array
-      * @param shift_y The y-offset of the top left corner of the shifted area
-      * wrt. to the y-index of the top left corner of the blitz::array.
-      * @param shift_x The x-offset of the top left corner of the shifted area 
-      * wrt. to the x-index of the top left corner of the blitz::array.
+      * @param shift_y The y-offset of the top left corner of the shifted 
+      *   area wrt. to the y-index of the top left corner of the 
+      *   blitz::array.
+      * @param shift_x The x-offset of the top left corner of the shifted 
+      *   area wrt. to the x-index of the top left corner of the 
+      *   blitz::array.
       * @param allow_out Whether an exception should be raised or not if the 
-      * shifted blitz::array has no pixel in common with the input blitz::array.
+      *   shifted blitz::array has no pixel in common with the input 
+      *   blitz::array.
       * @param zero_out Whether the shifted which is out of the boundary
-      * of the input blitz array should be filled with zero values, or with 
-      * the intensity of the closest pixel in the neighbourhood.
+      *   of the input blitz array should be filled with zero values, or 
+      *   with the intensity of the closest pixel in the neighbourhood.
       */
     template<typename T>
     void shift(const blitz::Array<T,2>& src, blitz::Array<T,2>& dst, 
-      const int shift_y, const int shift_x, const bool allow_out = false,
-      const bool zero_out = false)
+      const int shift_y, const int shift_x, 
+      const bool allow_out = false, const bool zero_out = false)
     {
       // Check parameters and throw exception if required
       if( !allow_out )
@@ -96,15 +101,18 @@ namespace bob {
       *   height (y-axis), whereas the third one is the width (x-axis).
       * @param src The input blitz array
       * @param dst The output blitz array
-      * @param shift_y The y-offset of the top left corner of the shifted area
-      * wrt. to the y-index of the top left corner of the blitz::array.
-      * @param shift_x The x-offset of the top left corner of the shifted area 
-      * wrt. to the x-index of the top left corner of the blitz::array.
+      * @param shift_y The y-offset of the top left corner of the shifted
+      *   area wrt. to the y-index of the top left corner of the 
+      *   blitz::array.
+      * @param shift_x The x-offset of the top left corner of the shifted
+      *   area wrt. to the x-index of the top left corner of the 
+      *   blitz::array.
       * @param allow_out Whether an exception should be raised or not if the 
-      * shifted blitz::array has no pixel in common with the input blitz::array.
+      *   shifted blitz::array has no pixel in common with the input 
+      *   blitz::array.
       * @param zero_out Whether the shifted which is out of the boundary
-      * of the input blitz array should be filled with zero values, or with 
-      * the intensity of the closest pixel in the neighbourhood.
+      *   of the input blitz array should be filled with zero values, or 
+      *   with the intensity of the closest pixel in the neighbourhood.
       */
     template<typename T>
     void shift(const blitz::Array<T,3>& src, blitz::Array<T,3>& dst, 
@@ -130,7 +138,8 @@ namespace bob {
         // Shift the 2D array
         blitz::Array<bool,2> src_mask, dst_mask; 
         detail::cropNoCheck<T,false>(src_slice, src_mask, dst_slice, 
-          dst_mask, shift_y, shift_x, src.extent(1), src.extent(2), zero_out);
+          dst_mask, shift_y, shift_x, src.extent(1), src.extent(2), 
+          zero_out);
       }
     }
 
@@ -146,21 +155,24 @@ namespace bob {
       * @param dst The output blitz array
       * @param dst_mask The output blitz array mask, specifying the valid
       *   pixels of dst.
-      * @param shift_y The y-offset of the top left corner of the shifted area
-      * wrt. to the y-index of the top left corner of the blitz::array.
-      * @param shift_x The x-offset of the top left corner of the shifted area 
-      * wrt. to the x-index of the top left corner of the blitz::array.
-      * @param allow_out Whether an exception should be raised or not if the 
-      * shifted blitz::array has no pixel in common with the input blitz::array.
+      * @param shift_y The y-offset of the top left corner of the shifted
+      *   area wrt. to the y-index of the top left corner of the 
+      *   blitz::array.
+      * @param shift_x The x-offset of the top left corner of the shifted 
+      *   area wrt. to the x-index of the top left corner of the 
+      *   blitz::array.
+      * @param allow_out Whether an exception should be raised or not if 
+      *   the shifted blitz::array has no pixel in common with the input 
+      *   blitz::array.
       * @param zero_out Whether the shifted which is out of the boundary
-      * of the input blitz array should be filled with zero values, or with 
-      * the intensity of the closest pixel in the neighbourhood.
+      *   of the input blitz array should be filled with zero values, or 
+      *   with the intensity of the closest pixel in the neighbourhood.
       */
     template<typename T>
-    void shift(const blitz::Array<T,2>& src, const blitz::Array<bool,2>& src_mask,
-      blitz::Array<T,2>& dst, blitz::Array<bool,2>& dst_mask,
-      const int shift_y, const int shift_x, const bool allow_out = false,
-      const bool zero_out = false)
+    void shift(const blitz::Array<T,2>& src, 
+      const blitz::Array<bool,2>& src_mask, blitz::Array<T,2>& dst, 
+      blitz::Array<bool,2>& dst_mask, const int shift_y, const int shift_x, 
+      const bool allow_out = false, const bool zero_out = false)
     {
       // Check parameters and throw exception if required
       if( !allow_out )
@@ -193,15 +205,18 @@ namespace bob {
       * @param dst The output blitz array
       * @param dst_mask The output blitz array mask, specifying the valid
       *   pixels of dst.
-      * @param shift_y The y-offset of the top left corner of the shifted area
-      * wrt. to the y-index of the top left corner of the blitz::array.
-      * @param shift_x The x-offset of the top left corner of the shifted area 
-      * wrt. to the x-index of the top left corner of the blitz::array.
-      * @param allow_out Whether an exception should be raised or not if the 
-      * shifted blitz::array has no pixel in common with the input blitz::array.
+      * @param shift_y The y-offset of the top left corner of the shifted
+      *   area wrt. to the y-index of the top left corner of the 
+      *   blitz::array.
+      * @param shift_x The x-offset of the top left corner of the shifted 
+      *   area wrt. to the x-index of the top left corner of the 
+      *   blitz::array.
+      * @param allow_out Whether an exception should be raised or not if 
+      *   the shifted blitz::array has no pixel in common with the input 
+      *   blitz::array.
       * @param zero_out Whether the shifted which is out of the boundary
-      * of the input blitz array should be filled with zero values, or with 
-      * the intensity of the closest pixel in the neighbourhood.
+      *   of the input blitz array should be filled with zero values, or 
+      *   with the intensity of the closest pixel in the neighbourhood. 
       */
     template<typename T>
     void shift(const blitz::Array<T,3>& src, const blitz::Array<bool,3>& src_mask,
